requirements.txt
django
gunicorn
h5py
matplotlib
mysqlclient
numpy
Pillow
scikit-learn
torch==1.4
torchvision==0.5.0